chore: move general funcs to utils
